block landscape to all pages #168
